--- a/tests/expectations/lib.rs
+++ b/tests/expectations/lib.rs
@@ -1,0 +1,3 @@
+#![allow(dead_code)]
+#![allow(non_camel_case_types)]
+#![allow(non_upper_case_globals)]