… src/codegen/mod.rs
… src/ir/context.rs
… src/ir/function.rs
… src/ir/item.rs
… src/ir/item_kind.rs
… src/ir/layout.rs
… src/ir/module.rs
… src/ir/named.rs
… src/ir/template.rs
… src/ir/ty.rs
… src/ir/var.rs
… src/parse.rs
… src/uses.rs
… tests/expectations/tests/anon_enum_trait.rs
… tests/expectations/tests/anon_union.rs
… tests/expectations/tests/auto.rs
… tests/expectations/tests/bad-namespace-parenthood-inheritance.rs
… tests/expectations/tests/class_nested.rs
… tests/expectations/tests/class_with_dtor.rs
… tests/expectations/tests/constant-non-specialized-tp.rs
… tests/expectations/tests/constructor-tp.rs
… tests/expectations/tests/crtp.rs
… tests/expectations/tests/dash_language.rs
… tests/expectations/tests/empty_template_param_name.rs
… tests/expectations/tests/enum_in_template_with_typedef.rs
… tests/expectations/tests/eval-variadic-template-parameter.rs
… tests/expectations/tests/forward-inherit-struct.rs
… tests/expectations/tests/in_class_typedef.rs
… tests/expectations/tests/inherit-namespaced.rs
… tests/expectations/tests/inherit_named.rs