The major changes in this PR are:

* De-duplication of named template type parameters (this is probably the biggest part of the PR, and nastiest because it is the part that deals with libclang)

* Removing the `signature_contains_named_type` stuff, and enabling the more sound template type parameter usage analysis that has been landing in bits here and there.

* **LOTS** of new tests for template type parameter usage

@emilio: can you also test this on the stylo bindings? I tested on the SpiderMonkey bindings and found a bug and fixed a bug related to instantiations of partially specialized templates. I'd like to make sure that there aren't any latent bugs uncovered in Stylo.

This is NOT ready to merge quite yet, but is ready for some more eyeballs that are not mine.

Still TODO:

* [x] Rebase so that these changes will merge cleanly -- I'll get on this ASAP

* [x] Time SpiderMonkey bindings generation with and without these changes to see what the overhead of the new analysis is (if any!) on Large and Real World bindings

* [x] Test theses changes on Stylo (thanks @emilio!)

    * [ ] (optional) and time Stylo bindings generation with and without these changes as well (only if you want to, @emilio)

Thanks!