@fitzgen 2017-02-27
(It the lack of seeing that the constructor is inline, but we need the constructor to trigger the panic that I was reducing when I generated this test case)