@fitzgen 2017-02-27
# Timing SpiderMonkey Bindings Generation

time | master | with these changes
---- | ------ | ------------------
real | 9.50   | 9.24
user | 9.35   | 9.10
sys  | 0.11   | 0.10