@fitzgen 2017-02-27
So it looks like we are spending about the same amount of time on bindings generation. If I keep regenerating bindings, the numbers fluctuate a bit, but are about the same in either case. Sometimes master is faster than with this PR, sometimes the other way around. Doesn't look like this PR has much of an impact on performance either way.