@fitzgen 2017-02-27
r? @emilio :)

This is ready for a real review now (assumign that you can use it to generate stylo bindings successfully)