@fitzgen 2017-02-27
> My main concern is that the actual unused template parameter can be pretty useful when writing abstractions on top of the bindgen-generated code.

This is a bit unfortunate. However, it is also the kind of thing where leveraging these templates on the Rust side already requires a bit of wrapping in a Rust-y API, and so a little bit more of that wrapping (adding a phantom type and whatever methods) doesn't seem like the end of the world. The trade off we're getting is more correctness, which almost always seems like a good trade off.

> It seems (from the reviewed code experience, which might not be so much) that switching from this code generation to the alternative (what we have now, template params + PhantomData) shouldn't be extremely hard (of course, it'd bring back the issues we have with template aliases). Is that right?

We could add an annotation that is like "please consider all my template parameters used" and it should be pretty easy. Unless it is crucial for Stylo, I'd like to do such things as a follow up.