@fitzgen 2017-02-27
> We could add an annotation that is like "please consider all my template parameters used" and it should be pretty easy. Unless it is crucial for Stylo, I'd like to do such things as a follow up.

And we could deny this annotation/flag/option on type aliases, which is where the correctness issues come into play.