@bors-servo 2017-03-01
:umbrella: The latest upstream changes (presumably #549) made this pull request unmergeable. Please resolve the merge conflicts.