@fitzgen 2017-03-01
For posterity, here is the `creduce`d version of the failing stylo bindings:

```c++
template <typename> class a;
template <typename b, typename... c> class a<b(c...)> { a(const a &); };
template <typename b, typename... c> a<b(c...)>::a(const a &) {}
```

I really hope that this isn't `creduce` "kindly" figuring out that we don't support anonymous template parameters again, and skipping past the real issue I was trying to pin down... Although I do have a branch with a fix for the anonymous type parameters that I should be able to rebase pretty quickly.