@fitzgen 2017-03-03
Small update:

I fixed that test case, but am `creduce`ing another failure when generating Stylo bindings.

I think the fully pre-processed Stylo bindings are even bigger than SpiderMonkeys, and `creduce` is taking almost 24h to come up with reductions on Stylo bindings :-/

I guess the bright side is that it's good to get more edge cases into our test suite so we won't have to do this as often in the future!