@fitzgen 2017-03-06
Next stylo-derived creduced test case:

```c++
template <typename a> struct b {
  template <typename> using c = typename a::d;
  template <typename e> using f = b<c<e>>;
  f<int> g;
};
```