@fitzgen 2017-03-06
Grr, we're generating infinitely sized types for the most recent creduced test case.