@fitzgen 2017-03-06
@emilio OK, this should be ready for final review now.

Fixing that last test case was pretty nasty. We can't do anything sane for associated types other than recognize them and treat them as opaque, but of course clang doesn't expose the type, so we have to do some nasty regex matching on the type's spelling :(