@emilio 2017-03-07
r=me. The regex thing makes me a bit uneasy, but not that I have a better solution, so... Please squash everything or remove the "Fix bad rebase" commit before landing :)