@fitzgen 2017-03-07
Fixed final nitpicks and squashed. Thanks for the review!

@bors-servo r=emilio