@bors-servo 2017-03-07
:pushpin: Commit 1c4332a has been approved by `emilio`

<!-- @bors-servo r=emilio 1c4332a1aa3dd299c133f72efb8b28dd5c8aa9d4 -->