@bors-servo 2017-03-07
:sunny: Test successful - [status-travis](https://travis-ci.org/servo/rust-bindgen/builds/208704560)
Approved by: emilio
Pushing dd1dc9e391f59e864316f145d1d3aac9294fd255 to master...